Move forkChoiceRule logic and update failing tests
