+++ framework/src/modules/chain/bft/bft.js
@@ -20,6 +20,7 @@ const {
 	EVENT_BFT_FINALIZED_HEIGHT_CHANGED,
 	FinalityManager,
 } = require('./finality_manager');
+const forkChoiceRule = require('./fork_choice_rule');
 const { validateBlockHeader } = require('./utils');
 
 const META_KEYS = {
@@ -49,12 +50,13 @@ class BFT extends EventEmitter {
 	 * @param {integer} activeDelegates - Number of delegates
 	 * @param {integer} startingHeight - The height at which BFT finalization manager initialize
 	 */
-	constructor({ storage, logger, activeDelegates, startingHeight }) {
+	constructor({ storage, logger, slots, activeDelegates, startingHeight }) {
 		super();
 		this.finalityManager = null;
 
 		this.logger = logger;
 		this.storage = storage;
+		this.slots = slots;
 		this.constants = {
 			activeDelegates,
 			startingHeight,
@@ -223,6 +225,51 @@ class BFT extends EventEmitter {
 			META_KEYS.LAST_BLOCK_FORGED,
 			previouslyForgedStr,
 		);
+	}
+
+	forkChoice({ lastBlock, block }) {
+		// Current time since Lisk Epoch
+		block.receivedAt = this.slots.getEpochTime();
+		// Cases are numbered following LIP-0014 Fork choice rule.
+		// See: https://github.com/LiskHQ/lips/blob/master/proposals/lip-0014.md#applying-blocks-according-to-fork-choice-rule
+		// Case 2 and 1 have flipped execution order for better readability. Behavior is still the same
+
+		if (forkChoiceRule.isValidBlock(lastBlock, block)) {
+			// Case 2: correct block received
+			return forkChoiceRule.FORK_STATUS_VALID_BLOCK;
+		}
+
+		if (forkChoiceRule.isIdenticalBlock(lastBlock, block)) {
+			// Case 1: same block received twice
+			return forkChoiceRule.FORK_STATUS_IDENTICAL_BLOCK;
+		}
+
+		if (forkChoiceRule.isDoubleForging(lastBlock, block)) {
+			// Delegates are the same
+			// Case 3: double forging different blocks in the same slot.
+			// Last Block stands.
+			return forkChoiceRule.FORK_STATUS_DOUBLE_FORGING;
+		}
+
+		if (
+			forkChoiceRule.isTieBreak({
+				slots: this.slots,
+				lastAppliedBlock: lastBlock,
+				receivedBlock: block,
+			})
+		) {
+			// Two competing blocks by different delegates at the same height.
+			// Case 4: Tie break
+			return forkChoiceRule.FORK_STATUS_TIE_BREAK;
+		}
+
+		if (forkChoiceRule.isDifferentChain(lastBlock, block)) {
+			// Case 5: received block has priority. Move to a different chain.
+			return forkChoiceRule.FORK_STATUS_DIFFERENT_CHAIN;
+		}
+
+		// Discard newly received block
+		return forkChoiceRule.FORK_STATUS_DISCARD;
 	}
 
 	async _getPreviouslyForgedMap() {
