+++ framework/src/modules/chain/block_processor_v2.js
@@ -221,7 +221,7 @@ class BlockProcessorV2 extends BaseBlockProcessor {
 		]);
 
 		this.forkStatus.pipe([
-			({ block, lastBlock }) => this.blocksModule.forkChoice(block, lastBlock), // validate common block header
+			({ block, lastBlock }) => this.bftModule.forkChoice(block, lastBlock), // validate common block header
 		]);
 
 		this.verify.pipe([({ block }) => this.bftModule.verifyNewBlock(block)]);
