+++ framework/src/modules/chain/blocks/blocks.js
@@ -43,7 +43,6 @@ const {
 	calculateReward,
 	calculateMilestone,
 } = require('./block_reward');
-const forkChoiceRule = require('../bft/fork_choice_rule');
 const {
 	validateSignature,
 	validatePreviousBlockProperty,
@@ -233,51 +232,6 @@ class Blocks extends EventEmitter {
 	async verifyInMemory(block, lastBlock) {
 		verifyPreviousBlockId(block, lastBlock, this.genesisBlock);
 		validateBlockSlot(block, lastBlock, this.slots);
-	}
-
-	forkChoice(block, lastBlock) {
-		// Current time since Lisk Epoch
-		block.receivedAt = this.slots.getEpochTime();
-		// Cases are numbered following LIP-0014 Fork choice rule.
-		// See: https://github.com/LiskHQ/lips/blob/master/proposals/lip-0014.md#applying-blocks-according-to-fork-choice-rule
-		// Case 2 and 1 have flipped execution order for better readability. Behavior is still the same
-
-		if (forkChoiceRule.isValidBlock(lastBlock, block)) {
-			// Case 2: correct block received
-			return forkChoiceRule.FORK_STATUS_VALID_BLOCK;
-		}
-
-		if (forkChoiceRule.isIdenticalBlock(lastBlock, block)) {
-			// Case 1: same block received twice
-			return forkChoiceRule.FORK_STATUS_IDENTICAL_BLOCK;
-		}
-
-		if (forkChoiceRule.isDoubleForging(lastBlock, block)) {
-			// Delegates are the same
-			// Case 3: double forging different blocks in the same slot.
-			// Last Block stands.
-			return forkChoiceRule.FORK_STATUS_DOUBLE_FORGING;
-		}
-
-		if (
-			forkChoiceRule.isTieBreak({
-				slots: this.slots,
-				lastAppliedBlock: lastBlock,
-				receivedBlock: block,
-			})
-		) {
-			// Two competing blocks by different delegates at the same height.
-			// Case 4: Tie break
-			return forkChoiceRule.FORK_STATUS_TIE_BREAK;
-		}
-
-		if (forkChoiceRule.isDifferentChain(lastBlock, block)) {
-			// Case 5: received block has priority. Move to a different chain.
-			return forkChoiceRule.FORK_STATUS_DIFFERENT_CHAIN;
-		}
-
-		// Discard newly received block
-		return forkChoiceRule.FORK_STATUS_DISCARD;
 	}
 
 	async verify(blockInstance, stateStore, { skipExistingCheck }) {
